requirements.txt
djangorestframework-simplejwt
djangorestframework
pymysql